Enum oxc_ast::ast::Expression

pub enum Expression<'a> {

    BooleanLiteral(Box<'a, BooleanLiteral>),
    NullLiteral(Box<'a, NullLiteral>),
    NumberLiteral(Box<'a, NumberLiteral<'a>>),
    BigintLiteral(Box<'a, BigintLiteral>),
    RegExpLiteral(Box<'a, RegExpLiteral>),
    StringLiteral(Box<'a, StringLiteral>),
    TemplateLiteral(Box<'a, TemplateLiteral<'a>>),
    Identifier(Box<'a, IdentifierReference>),
    MetaProperty(Box<'a, MetaProperty>),
    Super(Box<'a, Super>),
    ArrayExpression(Box<'a, ArrayExpression<'a>>),
    ArrowFunctionExpression(Box<'a, ArrowExpression<'a>>),
    AssignmentExpression(Box<'a, AssignmentExpression<'a>>),
    AwaitExpression(Box<'a, AwaitExpression<'a>>),
    BinaryExpression(Box<'a, BinaryExpression<'a>>),
    CallExpression(Box<'a, CallExpression<'a>>),
    ChainExpression(Box<'a, ChainExpression<'a>>),
    ClassExpression(Box<'a, Class<'a>>),
    ConditionalExpression(Box<'a, ConditionalExpression<'a>>),
    FunctionExpression(Box<'a, Function<'a>>),
    ImportExpression(Box<'a, ImportExpression<'a>>),
    LogicalExpression(Box<'a, LogicalExpression<'a>>),
    MemberExpression(Box<'a, MemberExpression<'a>>),
    NewExpression(Box<'a, NewExpression<'a>>),
    ObjectExpression(Box<'a, ObjectExpression<'a>>),
    ParenthesizedExpression(Box<'a, ParenthesizedExpression<'a>>),
    SequenceExpression(Box<'a, SequenceExpression<'a>>),
    TaggedTemplateExpression(Box<'a, TaggedTemplateExpression<'a>>),
    ThisExpression(Box<'a, ThisExpression>),
    UnaryExpression(Box<'a, UnaryExpression<'a>>),
    UpdateExpression(Box<'a, UpdateExpression<'a>>),
    YieldExpression(Box<'a, YieldExpression<'a>>),
    PrivateInExpression(Box<'a, PrivateInExpression<'a>>),
    JSXElement(Box<'a, JSXElement<'a>>),
    JSXFragment(Box<'a, JSXFragment<'a>>),
    TSAsExpression(Box<'a, TSAsExpression<'a>>),
    TSTypeAssertion(Box<'a, TSTypeAssertion<'a>>),
    TSNonNullExpression(Box<'a, TSNonNullExpression<'a>>),
    TSInstantiationExpression(Box<'a, TSInstantiationExpression<'a>>),

}

Section 13 Expression

Variants

BooleanLiteral(Box<'a, BooleanLiteral>)

NullLiteral(Box<'a, NullLiteral>)

NumberLiteral(Box<'a, NumberLiteral<'a>>)

BigintLiteral(Box<'a, BigintLiteral>)

RegExpLiteral(Box<'a, RegExpLiteral>)

StringLiteral(Box<'a, StringLiteral>)

TemplateLiteral(Box<'a, TemplateLiteral<'a>>)

Identifier(Box<'a, IdentifierReference>)

MetaProperty(Box<'a, MetaProperty>)

Super(Box<'a, Super>)

ArrayExpression(Box<'a, ArrayExpression<'a>>)

ArrowFunctionExpression(Box<'a, ArrowExpression<'a>>)

AssignmentExpression(Box<'a, AssignmentExpression<'a>>)

AwaitExpression(Box<'a, AwaitExpression<'a>>)

BinaryExpression(Box<'a, BinaryExpression<'a>>)

CallExpression(Box<'a, CallExpression<'a>>)

ChainExpression(Box<'a, ChainExpression<'a>>)

ClassExpression(Box<'a, Class<'a>>)

ConditionalExpression(Box<'a, ConditionalExpression<'a>>)

FunctionExpression(Box<'a, Function<'a>>)

ImportExpression(Box<'a, ImportExpression<'a>>)

LogicalExpression(Box<'a, LogicalExpression<'a>>)

MemberExpression(Box<'a, MemberExpression<'a>>)

NewExpression(Box<'a, NewExpression<'a>>)

ObjectExpression(Box<'a, ObjectExpression<'a>>)

ParenthesizedExpression(Box<'a, ParenthesizedExpression<'a>>)

SequenceExpression(Box<'a, SequenceExpression<'a>>)

TaggedTemplateExpression(Box<'a, TaggedTemplateExpression<'a>>)

ThisExpression(Box<'a, ThisExpression>)

UnaryExpression(Box<'a, UnaryExpression<'a>>)

UpdateExpression(Box<'a, UpdateExpression<'a>>)

YieldExpression(Box<'a, YieldExpression<'a>>)

PrivateInExpression(Box<'a, PrivateInExpression<'a>>)

JSXElement(Box<'a, JSXElement<'a>>)

JSXFragment(Box<'a, JSXFragment<'a>>)

TSAsExpression(Box<'a, TSAsExpression<'a>>)

TSTypeAssertion(Box<'a, TSTypeAssertion<'a>>)

TSNonNullExpression(Box<'a, TSNonNullExpression<'a>>)

TSInstantiationExpression(Box<'a, TSInstantiationExpression<'a>>)

Implementations

impl<'a> Expression<'a>

pub const fn is_primary_expression(&self) -> bool

PrimaryExpression tc39/ecma262#prod-PrimaryExpression

pub const fn is_literal_expression(&self) -> bool

pub const fn is_string_literal(&self) -> bool

pub const fn is_null(&self) -> bool

Determines whether the given expr is a null literal

pub fn is_undefined(&self) -> bool

Determines whether the given expr is a undefined literal

pub fn without_parenthesized(&self) -> &Self

Remove nested parentheses from this expression.

pub fn is_specific_id(&self, name: &str) -> bool

pub fn get_inner_expression(&self) -> &Expression<'a>

pub fn get_identifier_reference(&self) -> Option<&IdentifierReference>